requirements.txt
cffi==1.14.3
click==7.1.2
elasticsearch==7.6.0
Flask==1.1.2
graphite-api==1.1.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
structlog==20.1.0
taospy
tzlocal==2.1
urllib3==1.25.10
uWSGI==2.0.19.1
Werkzeug==1.0.1
